cinema_admin/requirements/base.txt
django==3.1
django-model-utils==4.0.0
psycopg2==2.8.6
python-dotenv==0.15.0
